requirements.txt
Flask>=2.3.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
numpy>=1.24.0
Pillow>=9.5.0
scipy>=1.10.0
Werkzeug>=2.3.0